modified planner page, added task blocks and ics export, parse task input implementation
Modified planner page to include suggested task blocks (3-5 of them) allowing users to swap the suggested times if it doesn't fit them. Implemented export to .ics functionality as well as auto parsing assist so that if the user doesn't want to add the task manually, they could type it into the textbox directly for it to be auto parsed into a task.